backend/requirements.txt
fastapi==0.111.1
uvicorn==0.30.3
sqlalchemy==2.0.31
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.0.1
black==24.4.2
pytest==8.3.2
pytest-asyncio==0.23.8
httpx==0.27.0
pillow==10.4.0